app/requirements.txt
flask == 1.0.2
mysql-connector-python == 8.0.25
python-dotenv == 0.10.1
flask-cors == 3.0.7
Flask-JWT == 0.3.2
gunicorn == 19.9.0
bcrypt == 3.1.6
